backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
sqlalchemy==2.0.27
aiosqlite==0.19.0
openai==1.12.0
pillow==10.2.0
imagehash==4.3.1
requests==2.31.0
aiohttp==3.9.3
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
faiss-cpu==1.7.4
numpy==1.26.4
httpx==0.26.0
razorpay==1.4.1
google-search-results==2.4.2
tiktoken==0.6.0
pytest==7.4.4
pytest-asyncio==0.21.1
# Optional: python-magic (requires libmagic system library)
# On macOS: brew install libmagic
# On Ubuntu: apt-get install libmagic1
